scummvm: fix floating dependencies
warning was:

WARNING: QA Issue: scummvm rdepends on fluidsynth, but it isn't a build dependency? [build-deps]
WARNING: QA Issue: scummvm rdepends on libflac, but it isn't a build dependency? [build-deps]
WARNING: QA Issue: scummvm rdepends on libfaad, but it isn't a build dependency? [build-deps]

fluidsynth is part of meta-multimedia - so disable it

Signed-off-by: Andreas Müller <schnitzeltony@googlemail.com>