supertux2: force build with tailored physfs
In preparation of 0.5.0 of supertux I sent supertux's version of physfs to
meta-oe. No I fund out, that other packages don't work with that version. So
I will send a patch to meta-oe to go back to old version of physfs.

To make supertux work again force it to use it's own version of physfs as
static library.

Signed-off-by: Andreas Müller <schnitzeltony@googlemail.com>